Add dedicated CSR model structure
