+++ src/Csr.XmlBuilder/CsrXmlBuilder.cs
@@ -293,81 +293,3 @@ public static class CsrXmlBuilder
         return mi.HasElements ? new XElement(gmi + "acquisitionInformation", mi) : null;
     }
 }
-
-public sealed class CsrExportBundle
-{
-    public string? CsrLocalId { get; set; }
-    public string CruiseName { get; set; } = string.Empty;
-    public string? Abstract { get; set; }
-    public DateTime? RevisionDateUtc { get; set; }
-    public DateTime? BeginUtc { get; set; }
-    public DateTime? EndUtc { get; set; }
-    public double? West { get; set; }
-    public double? East { get; set; }
-    public double? South { get; set; }
-    public double? North { get; set; }
-    public Organization? ResponsibleLab { get; set; }
-    public List<Keyword> SeaAreas { get; set; } = new();
-    public List<Mooring> Moorings { get; set; } = new();
-    public Distribution? Distribution { get; set; }
-    public DataQuality? DataQuality { get; set; }
-}
-
-public sealed class Organization
-{
-    public string Name { get; set; } = string.Empty;
-    public string? Url { get; set; }
-    public string? Country { get; set; }
-}
-
-public sealed class Keyword
-{
-    public string? Code { get; set; }
-    public string? Label { get; set; }
-}
-public sealed class Mooring
-{
-    public string? DataCategoryCode { get; set; }
-    public string? DataCategoryLabel { get; set; }
-    public string? Description { get; set; }
-    public double? Latitude { get; set; }
-    public double? Longitude { get; set; }
-    public DateTime? EventTimeUtc { get; set; }
-    public string? PrincipalInvestigator { get; set; }
-    public string? OrganizationName { get; set; }
-    public string? Email { get; set; }
-    public string? PlatformDescription { get; set; }
-}
-public sealed class Distribution
-{
-    public List<Format> Formats { get; set; } = new();
-    public List<OnlineResource> OnlineResources { get; set; } = new();
-    public string? UseLimitation { get; set; }
-    public string? AccessConstraints { get; set; }
-}
-
-public sealed class Format
-{
-    public string Name { get; set; } = string.Empty;
-    public string? Version { get; set; }
-}
-
-public sealed class OnlineResource
-{
-    public string Url { get; set; } = string.Empty;
-    public string? Description { get; set; }
-    public string? Protocol { get; set; }
-}
-
-public sealed class DataQuality
-{
-    public string? Lineage { get; set; }
-    public List<ProcessStep> ProcessSteps { get; set; } = new();
-}
-
-public sealed class ProcessStep
-{
-    public string? Description { get; set; }
-    public DateTime? Date { get; set; }
-    public Organization? Processor { get; set; }
-}
